Cambio de estilo/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
flask>=2.2.0
flask-cors>=3.0.10
Pillow>=9.5.0
numpy>=1.23.0
tqdm>=4.65.0